delete restaurant
