+++ api_gateway/src/main/kotlin/org/thechance/api_gateway/endpoints/RestaurantRoute.kt
@@ -21,7 +21,6 @@ fun Route.restaurantRoutes() {
 
     val restaurantGateway: IRestaurantGateway by inject()
     val identityGateway: IdentityGateway by inject()
-    val webSocketServerHandler: WebSocketServerHandler by inject()
 
     route("/restaurants") {
         get {
@@ -83,8 +82,15 @@ fun Route.restaurantRoutes() {
                     )
                 respondWithResult(HttpStatusCode.Created, newRestaurant)
             }
-        }
 
+            delete("/{restaurantId}") {
+                val (language, countryCode) = extractLocalizationHeader()
+                val restaurantId = call.parameters["restaurantId"]?.trim().toString()
+                val result = restaurantGateway.deleteRestaurant(restaurantId, Locale(language, countryCode))
+                respondWithResult(HttpStatusCode.Created, result)
+
+            }
+        }
 
         authenticateWithRole(Role.DASHBOARD_ADMIN) {
             put {
@@ -106,90 +112,12 @@ fun Route.restaurantRoutes() {
                 val updatedRestaurant = restaurantGateway.updateRestaurant(
                     locale = Locale(language, countryCode),
                     isAdmin = false,
-                    restaurant = restaurant.copy(ownerId =ownerId )
+                    restaurant = restaurant.copy(ownerId = ownerId)
                 )
                 respondWithResult(HttpStatusCode.OK, updatedRestaurant)
             }
         }
 
-        authenticateWithRole(Role.RESTAURANT_OWNER) {
 
-            route("/orders") {
-                webSocket("/{restaurantId}") {
-                    val restaurantId = call.parameters["restaurantId"]?.trim().orEmpty()
-                    val (language, countryCode) = extractLocalizationHeaderFromWebSocket()
-                    val orders = restaurantGateway.restaurantOrders(restaurantId, Locale(language, countryCode))
-                    webSocketServerHandler.sessions[restaurantId] = this
-                    webSocketServerHandler.sessions[restaurantId]?.let {
-                        webSocketServerHandler.tryToCollectFormWebSocket(
-                            orders,
-                            it
-                        )
-                    }
-                }
-
-                get("/{restaurantId}") {
-                    val restaurantId = call.parameters["restaurantId"]?.trim().orEmpty()
-                    val (language, countryCode) = extractLocalizationHeader()
-                    val result = restaurantGateway.getActiveOrders(restaurantId, Locale(language, countryCode))
-                    respondWithResult(HttpStatusCode.OK, result)
-                }
-
-
-                get("/history/{id}") {
-                    val id = call.parameters["id"]?.trim().toString()
-                    val page = call.parameters["page"]?.trim()?.toInt() ?: 1
-                    val limit = call.parameters["limit"]?.trim()?.toInt() ?: 10
-                    val (language, countryCode) = extractLocalizationHeader()
-                    val result = restaurantGateway.getOrdersHistory(
-                        restaurantId = id,
-                        page = page,
-                        limit = limit,
-                        locale = Locale(language, countryCode)
-                    )
-                    respondWithResult(HttpStatusCode.OK, result)
-                }
-
-                get("/count-by-days-back") {
-                    val id = call.parameters["restaurantId"]?.trim().toString()
-                    val daysBack = call.parameters["daysBack"]?.trim()?.toInt() ?: 7
-                    val (language, countryCode) = extractLocalizationHeader()
-                    val result = restaurantGateway.getOrdersCountByDaysBefore(
-                        restaurantId = id,
-                        daysBack = daysBack,
-                        locale = Locale(language, countryCode)
-                    )
-                    respondWithResult(HttpStatusCode.OK, result)
-                }
-
-                get("/revenue-by-days-back") {
-                    val id = call.parameters["restaurantId"]?.trim().toString()
-                    val daysBack = call.parameters["daysBack"]?.trim()?.toInt() ?: 7
-                    val (language, countryCode) = extractLocalizationHeader()
-                    val result = restaurantGateway.getOrdersRevenueByDaysBefore(
-                        restaurantId = id,
-                        daysBack = daysBack,
-                        locale = Locale(language, countryCode)
-                    )
-                    respondWithResult(HttpStatusCode.OK, result)
-                }
-
-                put("/{id}/status") {
-                    val id = call.parameters["id"]?.trim().toString()
-                    val params = call.receiveParameters()
-                    val status = params["status"]?.trim()?.toInt() ?: 0
-                    val (language, countryCode) = extractLocalizationHeader()
-                    val result = restaurantGateway.updateOrderStatus(id, status, Locale(language, countryCode))
-                    respondWithResult(HttpStatusCode.OK, result)
-                }
-
-                delete("/{restaurantId}") {
-                    val restaurantId = call.parameters["restaurantId"]?.trim().toString()
-                    val (language, countryCode) = extractLocalizationHeader()
-                    val result = restaurantGateway.deleteRestaurant(restaurantId, Locale(language, countryCode))
-                    respondWithResult(HttpStatusCode.OK, result)
-                }
-            }
-        }
     }
 }
