+++ api_gateway/src/main/kotlin/org/thechance/api_gateway/plugins/Routing.kt
@@ -18,5 +18,6 @@ fun Application.configureRouting(tokenConfiguration: TokenConfiguration) {
         cuisineRoute()
         taxiRoutes()
         mealRoute()
+        orderRoutes()
     }
 }
